Fix x64 VS2017 warnings
This are the last warnings in the ctp2 project VS2017 reports that do not come from automatically generated files. For those files Byacc and Flex would be needed to be updated or a newer version of them.

VS2017 does not report warnings for x84 SDL builds and two warnings for DirectX related code in DirectX x84 builds.

+++ ctp2_code/ctp/debugtools/debugassert.h
@@ -19,17 +19,17 @@ void DebugAssert_Close (void);
 
 #undef ASSERT
 #define ASSERT(condition)													\
-	{if (! ((int) (condition))) {											\
+	{if (! ((size_t) (condition))) {											\
 		DebugAssert_Assert ( __FILE__, __LINE__, LOG_ASSERT, #condition);	\
 	}}
 
 #define ASSERT_CLASS(log_class, condition)								\
-	{if (! ((int) (condition))) {										\
+	{if (! ((size_t) (condition))) {										\
 		DebugAssert_Assert (__FILE__, __LINE__, log_class, #condition);	\
 	}}
 
 #define ASSERT_INDIRECT(module_name, module_line, condition)					\
-	{if (! ((int) (condition))) {												\
+	{if (! ((size_t) (condition))) {												\
 		DebugAssert_Assert (module_name, module_line, LOG_ASSERT, #condition);	\
 	}}
